requirements.txt
open3d
numpy
scipy
matplotlib
pillow